+++ src/base/Link.php
@@ -8,13 +8,13 @@ use verbb\hyper\fieldlayoutelements\CustomAttributesField;
 use verbb\hyper\fieldlayoutelements\LinkField;
 use verbb\hyper\fieldlayoutelements\LinkTextField;
 use verbb\hyper\fieldlayoutelements\LinkTitleField;
+use verbb\hyper\helpers\Html;
 
 use Craft;
 use craft\base\Element;
 use craft\base\ElementInterface;
 use craft\helpers\App;
 use craft\helpers\ArrayHelper;
-use craft\helpers\Html;
 use craft\helpers\StringHelper;
 use craft\helpers\Template;
 use craft\models\FieldLayout;
@@ -24,6 +24,12 @@ use Twig\Markup;
 
 abstract class Link extends Element implements LinkInterface
 {
+    // Constants
+    // =========================================================================
+
+    public const SCENARIO_SETTINGS = 'settings';
+
+
     // Static Methods
     // =========================================================================
 
@@ -49,6 +55,11 @@ abstract class Link extends Element implements LinkInterface
         $linkTypeHandle = StringHelper::toPascalCase($context->label);
 
         return $context->field->handle . '_' . $linkTypeHandle . '_LinkType';
+    }
+
+    public static function checkElementUri(): bool
+    {
+        return false;
     }
 
     public static function getDefaultFieldLayout(): FieldLayout
@@ -176,14 +187,23 @@ abstract class Link extends Element implements LinkInterface
         return $this instanceof ElementLink;
     }
 
+    public function scenarios(): array
+    {
+        $scenarios = parent::scenarios();
+        $scenarios[self::SCENARIO_SETTINGS] = [];
+
+        return $scenarios;
+    }
+
     public function defineRules(): array
     {
         $rules = parent::defineRules();
 
-        $rules[] = [['label', 'handle'], 'required'];
+        // Validation for only when saving Hyper fields and their settings
+        $rules[] = [['label', 'handle'], 'required', 'on' => [self::SCENARIO_SETTINGS]];
 
         if ($this->isFieldRequired) {
-            $rules[] = [['linkValue'], 'required'];
+            $rules[] = [['linkValue'], 'required', 'on' => [self::SCENARIO_DEFAULT, self::SCENARIO_LIVE]];
         }
 
         return $rules;
@@ -511,6 +531,11 @@ abstract class Link extends Element implements LinkInterface
         // Ensure we don't have any duplicate classes
         if (isset($attributes['class']) && is_array($attributes['class'])) {
             $attributes['class'] = array_unique($attributes['class']);
+        }
+
+        // Handle `rel` attributes which have custom added support for merging
+        if (isset($attributes['rel']) && is_array($attributes['rel'])) {
+            $attributes['rel'] = implode(' ', $attributes['rel']);
         }
 
         return $attributes;