+++ src/fieldlayoutelements/AriaLabelField.php
@@ -16,6 +16,11 @@ class AriaLabelField extends TextField
 
     // Public Methods
     // =========================================================================
+    
+    public function showAttribute(): bool
+    {
+        return true;
+    }
 
     public function defaultLabel(?ElementInterface $element = null, bool $static = false): ?string
     {