+++ src/fieldlayoutelements/CustomAttributesField.php
@@ -17,6 +17,11 @@ class CustomAttributesField extends BaseNativeField
 
     // Public Methods
     // =========================================================================
+    
+    public function showAttribute(): bool
+    {
+        return true;
+    }
 
     public function defaultLabel(?ElementInterface $element = null, bool $static = false): ?string
     {