+++ src/links/Asset.php
@@ -14,4 +14,9 @@ class Asset extends ElementLink
     {
         return AssetElement::class;
     }
+
+    public static function checkElementUri(): bool
+    {
+        return false;
+    }
 }