+++ src/links/ShopifyProduct.php
@@ -3,20 +3,15 @@ namespace verbb\hyper\links;
 
 use verbb\hyper\base\ElementLink;
 
-use craft\commerce\elements\Variant as VariantElement;
+use craft\shopify\elements\Product as ShopifyProductElement;
 
-class Variant extends ElementLink
+class ShopifyProduct extends ElementLink
 {
     // Static Methods
     // =========================================================================
 
     public static function elementType(): string
     {
-        return VariantElement::class;
-    }
-
-    public static function checkElementUri(): bool
-    {
-        return false;
+        return ShopifyProductElement::class;
     }
 }